requirements.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==3.3.10
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bitarray==3.4.3
bitsandbytes==0.46.1
bleach==6.2.0
cbor2==5.6.5
certifi==2025.6.15
cffi==1.17.1
chardet @ file:///C:/ci_310/chardet_1642114080098/work
charset-normalizer==3.4.2
ckzg==2.1.1
codebase-to-text==1.0.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
crytic-compile==0.3.10
cycler==0.12.1
cytoolz==1.0.1
datasets==4.0.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.59.0
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
ipykernel==6.29.5
ipython==8.37.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kaggle==1.7.4.5
kiwisolver==1.4.8
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.4.4
notebook_shim==0.2.4
numpy==2.2.6
openai==1.3.0
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parsimonious==0.10.0
parso==0.8.4
pathlib2==2.3.7.post1
peft==0.16.0
pillow==11.3.0
platformdirs==4.3.8
prettytable==3.16.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
py-solc-x==2.0.4
pyarrow==20.0.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.5.0
pydantic_core==2.14.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
python-json-logger==3.3.0
python-slugify==8.0.4
pytz==2025.2
pyunormalize==16.0.0
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rlp==4.1.0
rpds-py==0.26.0
safetensors==0.5.3
Send2Trash==1.8.3
six==1.17.0
slither-analyzer==0.11.3
smmap==5.0.2
sniffio==1.3.1
solc-select==1.1.0
soupsieve==2.7
stack-data==0.6.3
sympy==1.14.0
terminado==0.18.1
text-unidecode==1.3
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.2
tomli==2.2.1
toolz==1.0.0
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.2
trl==0.19.1
types-python-dateutil==2.9.0.20250708
types-requests==2.32.4.20250611
typing_extensions==4.14.1
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.13
web3==7.12.0
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
